# Conflicts:
#	app/src/main/java/com/example/adam/myusefulllocations/Activity/MainActivity.java
#	app/src/main/java/com/example/adam/myusefulllocations/Fragment/ItemSearchFragment.java
#	app/src/main/java/com/example/adam/myusefulllocations/Fragment/MyItemRecyclerViewAdapter.java
#	app/src/main/res/layout/fragment_item.xml
#	app/src/main/res/layout/fragment_search_item_list.xml
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/com/example/adam/myusefulllocations/Activity/DataPassListener.java
@@ -2,6 +2,6 @@ package com.example.adam.myusefulllocations.Activity;
 
 public interface DataPassListener {
 
-    void passDataLocationToMap (float lat, float lng, String name);
+    void passDataLocationToMap (double lat, double lng, String name);
 
 }
